+++ src/copyToClipboard/copyToClipboard.stories.js
@@ -1,0 +1,16 @@
+import README from './copyToClipboard.md'
+
+export default {
+  title: 'Utils|copyToClipboard',
+  parameters: {
+    readme: {
+      content: README
+    }
+  }
+}
+
+export const Basic = () => {}
+
+Basic.story = {
+  name: 'Basic usage'
+}